+++ cmd/repositories.go
@@ -1,16 +1,14 @@
 package cmd
 
 import (
-	"n3dr/cli"
+	"github.com/030/n3dr/cli"
 
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
 var (
-	names  bool
-	count  bool
-	backup bool
+	count, names, backup bool
 )
 
 // repositoriesCmd represents the repositories command